+++ app/containers/MessageComposer/components/Quotes/index.ts
@@ -1,0 +1,1 @@
+export * from './Quotes';
